refactor: move hexagen modules into package
