+++ hexagen/hexagen.py
@@ -22,7 +22,7 @@ from scipy.spatial.transform import Rotation
 from typing import Callable, Optional, List  # Union
 
 from constants.constants import COLORS, WIDTH, HEIGHT, DIRECTIONS
-import src.plot_board as pb
+from . import plot_board as pb
 from utils.reading_tasks import read_task
 
 logger = logging.getLogger(__name__)
